project/backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyYAML
fastmcp